feat(server): implement message conversion and storage infrastructure
- Introduce AiMessageConverter, ChatMessageConverter, and ToolResultMessageConverter for converting between different message formats.
- Create CustomChatMemoryStore to manage chat message retrieval and storage, integrating with existing services.
- Define MessageWithBlocks record to encapsulate messages and their associated blocks.
- Extend MessageBlockEnum to include new content types: IMAGE, VIDEO, PDF, and AUDIO.
- Update AiMessageMapper and related services to work with CustomAiMessage for enhanced message handling.
- Implement UserMessageConverter for processing user messages with various content types.
- Refactor AiMessageService and AiMessageBlockService to support batch operations and improve message management.

+++ Data-Agent-Server/data-agent-server-app/src/main/java/edu/zsc/ai/domain/model/entity/ai/AiConversation.java
@@ -26,7 +26,7 @@ public class AiConversation {
 
     private Integer tokenCount;
 
-    private Short deleteFlag;
+    private Integer deleteFlag;
 
     private LocalDateTime createdAt;
